requirements.txt
aiodns==3.0.0; python_full_version >= "3.8.0" and python_version < "3.11"
aiofiles==0.8.0; python_version >= "3.6" and python_version < "4.0"
aiohttp==3.8.1; python_version >= "3.6"
aiosignal==1.2.0; python_full_version >= "3.8.0" and python_version < "3.11" and python_version >= "3.6"
alembic==1.8.0; python_version >= "3.7"
apscheduler==3.9.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0" and python_version < "4")
async-timeout==4.0.2; python_full_version >= "3.8.0" and python_version < "3.11" and python_version >= "3.6"
asyncpg==0.25.0; python_full_version >= "3.6.0"
attrs==21.4.0; python_full_version >= "3.8.0" and python_version < "3.11" and python_version >= "3.6"
brotli==1.0.9; python_full_version >= "3.8.0" and python_version < "3.11"
cachetools==5.2.0; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.10.1" and python_full_version < "3.11.0"
cchardet==2.1.7; python_full_version >= "3.8.0" and python_version < "3.11"
certifi==2022.6.15; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.10.1" and python_full_version < "3.11.0"
cffi==1.15.0; python_full_version >= "3.8.0" and python_version < "3.11"
charset-normalizer==2.0.12; python_version >= "3.7" and python_version < "3.11" and python_full_version >= "3.10.1" and python_full_version < "3.11.0"
ciso8601==2.2.0; python_full_version >= "3.8.0" and python_version < "3.11"
colorama==0.4.5; python_full_version >= "3.8.0" and python_version < "3.11" and python_version >= "3.6" and sys_platform == "win32"
colorlog==6.6.0; python_full_version >= "3.8.0" and python_version < "3.11" and python_version >= "3.6"
frozenlist==1.3.0; python_full_version >= "3.8.0" and python_version < "3.11" and python_version >= "3.7"
google-auth-oauthlib==0.5.2; python_full_version >= "3.10.1" and python_full_version < "3.11.0" and python_version >= "3.6"
google-auth==2.8.0; python_full_version >= "3.10.1" and python_full_version < "3.11.0" and python_version >= "3.6"
greenlet==1.1.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") and python_full_version >= "3.5.0"
gspread==5.4.0; python_full_version >= "3.10.1" and python_full_version < "3.11.0" and python_version >= "3.6"
hikari-lightbulb==2.2.2; python_full_version >= "3.8.0" and python_version < "3.11"
hikari==2.0.0.dev109; python_full_version >= "3.8.0" and python_version < "3.11"
idna==3.3; python_version >= "3.7" and python_version < "3.11" and python_full_version >= "3.10.1" and python_full_version < "3.11.0"
lightbulb-ext-wtf @ git+https://github.com/tandemdude/lightbulb-ext-wtf.git@master ; python_full_version >= "3.8.0" and python_version < "3.11"
mako==1.2.0; python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
multidict==6.0.2; python_full_version >= "3.8.0" and python_version < "3.11" and python_version >= "3.7"
oauthlib==3.2.0; python_full_version >= "3.10.1" and python_full_version < "3.11.0" and python_version >= "3.6"
postgres==4.0
psycopg2-binary==2.9.3; python_version >= "3.6"
psycopg2-pool==1.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
pyasn1-modules==0.2.8; python_full_version >= "3.10.1" and python_full_version < "3.11.0" and python_version >= "3.6"
pyasn1==0.4.8; python_full_version >= "3.10.1" and python_full_version < "3.11.0" and python_version >= "3.6" and python_version < "4"
pycares==4.2.0; python_full_version >= "3.8.0" and python_version < "3.11"
pycparser==2.21; python_full_version >= "3.8.0" and python_version < "3.11"
pytz-deprecation-shim==0.1.0.post0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
pytz==2022.1; python_full_version >= "3.10.1" and python_full_version < "3.11.0" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4")
requests-oauthlib==1.3.1; python_full_version >= "3.10.1" and python_full_version < "3.11.0" and python_version >= "3.6"
requests==2.28.0; python_version >= "3.7" and python_version < "4"
rsa==4.8; python_version >= "3.6" and python_version < "4" and python_full_version >= "3.10.1" and python_full_version < "3.11.0"
sector-accounting @ git+https://github.com/gs729/sector_accounting.git@b6059f4 ; python_full_version >= "3.10.1" and python_full_version < "3.11.0"
six==1.16.0; python_full_version >= "3.10.1" and python_version < "4" and python_full_version < "3.11.0" and python_version >= "3.6"
sqlalchemy==1.4.39; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
tzdata==2022.1; python_version >= "3.6" and python_full_version < "3.0.0" and platform_system == "Windows" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6" and platform_system == "Windows"
tzlocal==4.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.6"
urllib3==1.26.9; python_full_version >= "3.10.1" and python_version < "4" and python_version >= "3.7" and python_full_version < "3.11.0"
uvloop==0.16.0; python_version >= "3.7"
yarl==1.7.2; python_full_version >= "3.8.0" and python_version < "3.11" and python_version >= "3.6"
